convert localdb size from int to long

+++ server/src/main/java/password/pwm/http/servlet/configmanager/ConfigManagerWordlistServlet.java
@@ -241,7 +241,7 @@ public class ConfigManagerWordlistServlet extends AbstractPwmServlet
                         wordlistType.name() + "_wordCount",
                         DisplayElement.Type.number,
                         "Word Count",
-                        Integer.toString( wordlist.size() ) ) );
+                        Long.toString( wordlist.size() ) ) );
 
                 if ( wordlistStatus.isCompleted() )
                 {
