+++ server/src/main/java/password/pwm/svc/wordlist/Wordlist.java
@@ -33,7 +33,7 @@ import java.io.InputStream;
 public interface Wordlist extends PwmService
 {
 
-    int size( );
+    long size( );
 
     WordlistStatus readWordlistStatus( );
